revert back to component instead of hook
