+++ frontend/src/scenes/dashboard/Dashboard.tsx
@@ -10,7 +10,7 @@ import { CalendarOutlined, ReloadOutlined } from '@ant-design/icons'
 import moment from 'moment'
 import { Button } from 'antd'
 import './Dashboard.scss'
-import { useHotKeys } from 'lib/hooks/useHotKeys'
+import { KeyboardHotkeys } from 'lib/components/KeyboardHotkeys'
 
 interface Props {
     id: string
@@ -68,11 +68,10 @@ function DashboardView(): JSX.Element {
         )
     }
 
-    useHotKeys(HOTKEYS)
-
     return (
         <div className="dashboard">
             {dashboardMode !== 'public' && <DashboardHeader />}
+            <KeyboardHotkeys hotkeys={HOTKEYS} />
             {items && items.length ? (
                 <div>
                     <div className="dashboard-items-actions">
